Moved internal classes to impl package
