+++ xml/src/main/scala/akka/stream/alpakka/xml/javadsl/XmlWriting.scala
@@ -8,7 +8,7 @@ import java.nio.charset.{Charset, StandardCharsets}
 
 import akka.NotUsed
 import akka.stream.alpakka.xml.ParseEvent
-import akka.stream.alpakka.xml.Xml.StreamingXmlWriter
+import akka.stream.alpakka.xml.impl
 import akka.stream.scaladsl.Flow
 import akka.util.ByteString
 
@@ -19,13 +19,13 @@ object XmlWriting {
    * encoding UTF-8
    */
   def writer(): akka.stream.javadsl.Flow[ParseEvent, ByteString, NotUsed] =
-    Flow.fromGraph(new StreamingXmlWriter(StandardCharsets.UTF_8)).asJava
+    Flow.fromGraph(new impl.StreamingXmlWriter(StandardCharsets.UTF_8)).asJava
 
   /**
    * Writer Flow that takes a stream of XML events similar to SAX and write ByteStrings.
-   * @param strEncoding encoding of the stream
+   * @param charset encoding of the stream
    */
   def writer(charset: Charset): akka.stream.javadsl.Flow[ParseEvent, ByteString, NotUsed] =
-    Flow.fromGraph(new StreamingXmlWriter(charset)).asJava
+    Flow.fromGraph(new impl.StreamingXmlWriter(charset)).asJava
 
 }
